Distinguish between dotfile repos
The instructions could not be copied and pasted line by line.

The bare cloned repo shows up as `dotfiles.git` instead of `dotfiles`, I changed the command to `cd dotfiles.git` so it can be copy and pasted.

Also, it's confusing when both the source repository and the user's personal repo are named `dotfiles.git`. I changed the command to mirror repo to make this clear.